+++ benchmark/internal/logger/log.go
