+++ benchmark/serumapi/serumws.go
@@ -5,8 +5,8 @@ import (
 	"crypto/tls"
 	"encoding/json"
 	"fmt"
-	"github.com/bloXroute-Labs/serum-client-go/examples/benchmark/internal/arrival"
-	"github.com/bloXroute-Labs/serum-client-go/examples/benchmark/internal/logger"
+	"github.com/bloXroute-Labs/serum-client-go/benchmark/internal/arrival"
+	"github.com/bloXroute-Labs/serum-client-go/benchmark/internal/logger"
 	pb "github.com/bloXroute-Labs/serum-client-go/proto"
 	"github.com/gorilla/websocket"
 	"github.com/pkg/errors"
